Updated Delete Operation for Git (#2435)

+++ litmus-portal/graphql-server/pkg/gitops/handler/handler.go
@@ -205,6 +205,14 @@ func DeleteWorkflowFromGit(ctx context.Context, workflow *model.ChaosWorkFlowInp
 	}
 
 	workflowPath := gitops.ProjectDataPath + "/" + gitConfig.ProjectID + "/" + workflow.WorkflowName + ".yaml"
+	exists, err := gitops.PathExists(workflowPath)
+	if err != nil {
+		return errors.New("Cannot delete workflow from git : " + err.Error())
+	}
+	if !exists {
+		log.Print("File not found in git!")
+		return nil
+	}
 	err = os.RemoveAll(gitConfig.LocalPath + "/" + workflowPath)
 	if err != nil {
 		return errors.New("Cannot delete workflow from git : " + err.Error())
